Bug fix: birthday is today
When a birthday is today, the website doesn't show the date between brackets, but it shows the age of the person). To test this, we can split the string on a space and try to access the 2nd element; this will fail if there is only 1 number (the age). If it fails, we insert the current date to resolve this issue.